Remove copy of upstream Hanami::Action::CSRFProtection module
We use dry-inflector as inflector for Zeitwerk.

As of dry-inflector 0.2.0, it didn’t support CSRF as an acronym. Zeitwerk was
looking for `CsrfProtection`, so we had to temporarily add a module with
this exact inflection in this repository.

As of dry-inflector 0.2.1, it added that CSRF acronym.

That means now Zeitwerk is able to look for `CSRFProtection` module from
hanami-controller. As per this change, we can remove `CsrfProtection` from
the app template.