Update to database_cleaner-sequel 2.0
database_cleaner 2.0 ships with a separated adapter for sequel.

We update the database_cleaner configuration to adopt the new setup.

On top of that, previous to this commit, the configuration with
database_cleaner `~> 1.7` wasn't working with the SQLite adapter OOTB. That was
because database_cleaner expected `sqlite3` as the adapter name to
consider a connection as local and therefore allow it:

https://github.com/DatabaseCleaner/database_cleaner/pull/529/commits/7519d4381073dbcdda732a09d7c92a70fe312f8c

On the other hand, Sequel uses `sqlite` (without the `3`) as the adapter name:

https://github.com/jeremyevans/sequel/blob/2392145025c60d3c127bc1b921130598409f249c/lib/sequel/database/connecting.rb#L11

To work around that, one needed to disable [database_cleaner
safeguards](https://github.com/DatabaseCleaner/database_cleaner#safeguards).
However, with the new adapter there's no need to do that.